requirements.txt
gradio
torch
torchvision
transformers
Pillow
requests